requirements.txt
openai>=1.30.0
aiohttp>=3.9.0
cryptography>=42.0.0
python-dotenv>=1.0.0
pydantic>=2.7.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dateutil>=2.9.0
tenacity>=8.3.0
eval_type_backport>=0.2.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
